% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisSetupElastic.R
\name{basisSetupElastic}
\alias{basisSetupElastic}
\title{Elastic FDA Basis Expansion Calculation}
\usage{
basisSetupElastic(
  Y,
  warpData = NULL,
  basisType = "jfpcah",
  nBasis = NULL,
  propVarExplained = 0.99,
  srvf = FALSE
)
}
\arguments{
\item{Y}{A matrix of dimension nxq, where n is the number of observations of q multivariate/functional variables}

\item{warpData}{`time_warping` object from `fdasrvf` if `basisType=jfpca` or `jfpcah`}

\item{basisType}{The type of basis functions to use. Options are `jfpca` or `jfpcah`.}

\item{nBasis}{An integer specifying the number of basis functions to use. The default is NULL, in which case propVarExplained is used to choose nBasis.}

\item{propVarExplained}{Proportion (between 0 and 1) of variation to explain when choosing the number of principal components. Only used if nBasis is NULL (which is the default).}

\item{srvf}{use SRVF if `basisType=jfpca` or `jfpcah` default = TRUE}
}
\value{
An object of class "basisSetup" containing information about the basis decomposition: "Y" is the original response matrix, "Ycenter" and "Yscale" are vectors of length q specifying the centering and scaling values used for each dimension of Y, "basisType" is the user-specified type of basis, the q-vector "varExplained" specifies the amount of variance explained by each basis in the expansion, matrix "basis" is the qxq basis matrix (e.g., the principal component vectors in the case of pca), "coefs" is the nxq matrix of observation-specific basis weights (e.g., the "scores" in the case of pca), and "truncError" is the nxq matrix of residuals, after accounting for the first nBasis bases.
}
\description{
Calculates a basis expansion of a functional data. Used by the mvBayes function to calculate a basis expansion of a response matrix.
}
\seealso{
Used by the main function \link{mvBayes}
}
