% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvBayes.R
\name{mvBayes}
\alias{mvBayes}
\title{Multivariate Bayesian Regression}
\usage{
mvBayes(
  bayesModel,
  X,
  Y,
  basisType = "pca",
  customBasis = NULL,
  nBasis = NULL,
  propVarExplained = 0.99,
  center = TRUE,
  scale = FALSE,
  nCores = 1,
  samplesExtract = NULL,
  residSDExtract = NULL,
  idxSamplesArg = "idxSamples",
  ...
)
}
\arguments{
\item{bayesModel}{A Bayesian regression model-fitting function, with first argument taking an nxp input matrix or data.frame, and second argument taking an n-vector of numeric responses.}

\item{X}{A matrix of predictors of dimension nxp, where n is the number of training examples and p is the number of inputs (features).}

\item{Y}{A response matrix of dimension nxq, where q is the number of multivariate/functional responses.}

\item{basisType}{The type of basis functions to use. Options are `pca`, `pns`, `splinet`, `bspline`, or `legendre`.}

\item{customBasis}{Optional user-provided basis of dimension `c(nBasis, nMV)`, only used if `basisType==custom`.}

\item{nBasis}{An integer specifying the number of basis functions to use. The default is NULL, in which case propVarExplained is used to choose nBasis.}

\item{propVarExplained}{Proportion (between 0 and 1) of variation to explain when choosing the number of principal components. Only used if nBasis is NULL (which is the default).}

\item{center}{A logical argument specifying whether or not to center the responses before computing the basis decomposition.}

\item{scale}{A logical argument specifying whether or not to scale by the standard deviation of each component of the response before computing the basis decomposition.}

\item{nCores}{An integer less than or equal to nBasis, specifying the number of threads to use when fitting independent Bayesian models.}

\item{samplesExtract}{function taking the output of bayesModel (`bm`) and extracting posterior samples of all parameters of interest. If `NULL`, mvBayes tries to access`bm$samples`; if unsuccessful, an object called `samples` is created with attribute `residSD`.}

\item{residSDExtract}{function taking the output of bayesModel (`bm`) and extracting posterior samples of the residual standard deviation (`residSD`). If `NULL``, mvBayes tries to access `bm$samples$residSD`; if unsuccessful, `residSD` is the  standard deviation of the residuals.}

\item{idxSamplesArg}{str Name of an optional argument of `predict` controlling which posterior samples are used for posterior prediction.}

\item{...}{Additional arguments to bayesModel.}
}
\value{
An object of class "mvBayes", which is a list containing "X", an object called "basisInfo" of class "basisSetup" containing information about the basis decomposition, "bayesModel", and "bmList", which contains a list of length nBasis containing fitted model objects for each basis component.
}
\description{
Wrapper to fit a multivariate response, using an arbitrary univariate Bayesian regression model to independently fit basis components (e.g., principal components) of the response.
}
\details{
First uses the basisSetup function to decompose the response into nBasis components, then independently fits bayesModel to each of those components.
}
\examples{
## simulate data (Friedman function with first variable as functional)
f = function(x) {
  10 * sin(pi * x[, 1] * x[, 2]) + 20 * (x[, 3] - .5)^2 + 10 * x[, 4] + 5 *
    x[, 5]
}
sigma = 1 # noise sd
n = 500 # number of observations
n_grid = 50 # size of functional variable grid
t_grid = seq(0, 1, length.out = n_grid) # functional grid
X = matrix(runif(n * 9), n, 9) # 9 non-functional variables, only first 4 matter
X_expanded = cbind(rep(t_grid, each = n), kronecker(rep(1, n_grid), X)) # to get y
Y = matrix(f(X_expanded), nrow = n) + rnorm(n * n_grid, 0, sigma)

## fit Multivariate BASS
fit = mvBayes(
  BASS::bass, X, Y, nBasis = 3, # mvBayes parameters
  samplesExtract = function(bm) list(
    s2 = bm$s2,
    nbasis = bm$nbasis,
    beta.prec = bm$beta.prec
  ),
  idxSamplesArg = 'mcmc.use'
)
}
\seealso{
\link{basisSetup} for computing the basis decomposition, \link{predict.mvBayes} for prediction, \link{plot.mvBayes} for plotting the model fit, \link{traceplot} for monitoring posterior convergence, and \link{mvSobol} for sensitivity analysis.
}
