% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{mvCV}
\alias{mvCV}
\title{Cross-Validation (CV) of a Multivariate Bayesian Regression Model}
\usage{
mvCV(
  bayesModel,
  X,
  Y,
  kFolds = NULL,
  nRep = 1,
  nTrain = NULL,
  nTest = NULL,
  seed = NULL,
  coverageTarget = 0.95,
  idxSamples = "all",
  uqTruncMethod = c("gaussian", "empirical"),
  warpData = NULL,
  ...
)
}
\arguments{
\item{bayesModel}{A Bayesian regression model-fitting function, with first argument taking an nxp input matrix or data.frame, and second argument taking an n-vector of numeric responses.}

\item{X}{A matrix of predictors of dimension nxp, where n is the total number of examples (including training and test sets) and p is the number of inputs (features).}

\item{Y}{A response matrix of dimension nxq, where q is the number of multivariate/functional responses.}

\item{kFolds}{Number of test sets to partition data. If kFolds=NULL (default), test sets are instead formed by random sample, via the nReps, nTrain, and nTest arguments.}

\item{nRep}{Number of repetitions of CV process. Only used when kFolds=NULL (default).}

\item{nTrain}{Number of examples to use in the training set. Only used when kFolds=NULL (default). If nTrain=NULL, nTrain is set to n - nTest; unless nTest is also NULL, in which case nTrain is set to ceiling(n/2).}

\item{nTest}{Number of examples to use in the test set. Only used when kFolds=NULL (default). If nTest=NULL, nTest is set to n - nTrain.}

\item{seed}{Randomization seed, for replication of the train/test split. The seed is un-initialized immediately after assigning the train/test split. If NULL, no seed is set.}

\item{coverageTarget}{level of coverage desired (default: 0.95)}

\item{idxSamples}{which samples to use in CV (default: "all")}

\item{uqTruncMethod}{method to use for UQ truncation (c("gaussian", "empirical"))}

\item{warpData}{`time_warping` object from `fdasrvf`. If supplied, \link{mvBayesElastic} is used instead of \link{mvBayes}, and each fold is aligned using the corresponding columns of `warpData`.}

\item{...}{Additional arguments to mvBayes, including arguments to bayesModel.}
}
\value{
An object of class "mvBayesCV", which is a list containing out-of-sample performance metrics for each replication, including rmse, rSquared, coverage, intervalWidth, intervalScore, crps, fitting and prediction times, and the function call.
}
\description{
Wrapper to fit and perform cross-validation for a multivariate Bayesian regression model, using the mvBayes function.
}
\details{
First separates the data into randomly chosen test and training sets (user-specified train/test splits and k-fold cv are forthcoming), then fits mvBayes(bayesModel, X, Y, ...) to the training set and evaluates predictive performance on the test set. Repeats this process nRep times.
}
\seealso{
\link{mvBayes}, \link{predict.mvBayes} for prediction
}
