% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol.R
\name{mvSobol}
\alias{mvSobol}
\title{Sobol' Indices for Multivariate Response}
\usage{
mvSobol(
  object,
  totalSobol = TRUE,
  idxSamples = "final",
  nMC = NULL,
  showPlot = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class "mvBayes" containing the multivariate Bayesian model fit for a specific basis expansion of a response matrix Y.}

\item{totalSobol}{whether to compute the total Sobol' index (summing all interactions)}

\item{idxSamples}{Use final MCMC sample}

\item{nMC}{Number of Monte Carlo iterations to compute, if None will attempt `anova` if available. Parameter is log2( n ), i.e., it makes 2^n points}

\item{showPlot}{whether to compute generate a plot along with the returned Sobol' object}

\item{...}{Additional arguments to BASS::sobolBasis (other emulator-specific Sobol' calculations to be implemented in the future). Unnecessary for anova-compatible models.}
}
\value{
An object of class "bassSob" if object$bayesModel is the BASS::bass() function, or "mvSobol" otherwise. Contains information about the Sobol' decomposition: See ?BASS::sobolBasis for more info.
}
\description{
Given an object of class "mvBayes" from the mvBayes() function, mvSobol() calculates the Sobol' indices for each input variable, assuming the response is multivariate. Currently only works if object$bayesModel is the BASS::bass() function, or if object$bayesModel is compatible with the anova() function.
}
\seealso{
See \link{mvBayes}
}
