% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvBayes.R
\name{plot.mvBayes}
\alias{plot.mvBayes}
\title{Plot the Bayesian Model Fit of a Multivariate Response}
\usage{
\method{plot}{mvBayes}(
  x,
  Xtest = NULL,
  Ytest = NULL,
  idxSamples = "final",
  nPlot = NULL,
  idxMV = NULL,
  xscale = "linear",
  xlabel = "Multivariate Index",
  title = NULL,
  file = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class "mvBayes" containing the multivariate Bayesian model fit of a response matrix Y.}

\item{Xtest}{A matrix or data.frame of inputs at which the model fit will be evaluated. If NULL, the training inputs X will be used.}

\item{Ytest}{A matrix of responses for which the model fit will be evaluated. Should correspond to Xtest. If NULL, the training responses Y will be used.}

\item{idxSamples}{str which samples to use}

\item{nPlot}{A positive integer specifying the number of samples to plot. Default is min(n, 1000), where n is nrow(Xtest) (or nrow(object$X) if Xtest is NULL).}

\item{idxMV}{A vector describing the sample points}

\item{xscale}{'linear' or 'log' of x-axis}

\item{xlabel}{str for x-axis label}

\item{title}{An optional title to be printed at the top of the diagnostic plots.}

\item{file}{An optional location at which the plot will be saved. If NULL, no file is saved.}

\item{...}{additional plot arguments}
}
\value{
Top left: Residuals after applying mvBayes(), plotted on top of Ytest (centered if centered=TRUE in the mvBayes() call). Top right: Residuals decomposed and colored by basis. Bottom left: R^2 values for each component. Bottom right: %Residual variance attribution for each component.
}
\description{
Given an object of class "mvBayes" from the mvBayes() function, plot.mvBayes() plots a few aspects of the Bayesian model fit, colored by basis.
}
\seealso{
See \link{mvBayes}, \link{predict.mvBayes}
}
