% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol.R
\name{plot.sobol}
\alias{plot.sobol}
\title{Plot Sobol Decomposition}
\usage{
\method{plot}{sobol}(
  x,
  totalSobol = TRUE,
  labels = NULL,
  idxMV = NULL,
  xscale = "linear",
  xlabel = "Multivariate Index",
  yOverlay = NULL,
  yOverlayLabel = "Overlay",
  waterfall = FALSE,
  title = NULL,
  file = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class "mvSobol" containing the Sobol Indices}

\item{totalSobol}{A boolean to plot the total sobol (default = `TRUE`)}

\item{labels}{A character vector of length <= 8 containing the names of the parameters}

\item{idxMV}{A vector defining the time points}

\item{xscale}{string whether to plot on a "linear" scale or "log"}

\item{xlabel}{string for the xlabel}

\item{yOverlay}{A boolean if to overlay the experimental function (default = `NULL`)}

\item{yOverlayLabel}{A string defining the overlay label}

\item{waterfall}{bool whether to plot sobol as a waterfall (functional pie-chart) default (`FALSE`)}

\item{title}{An optional title to be printed at the top of the traceplots.}

\item{file}{An optional location at which the plots will be saved. If NULL, no file is saved.}

\item{...}{additional plot arguments}
}
\description{
Given an object of class "mvSobol" from the mvSobol() function
}
\seealso{
See \link{mvBayes}
}
