% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvBayes.R
\name{traceplot.mvBayes}
\alias{traceplot.mvBayes}
\title{Traceplots from the Bayesian Model Fit of a Multivariate Response}
\usage{
\method{traceplot}{mvBayes}(
  object,
  modelParams = NULL,
  labels = NULL,
  title = NULL,
  file = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class "mvBayes" containing the multivariate Bayesian model fit of a response matrix Y.}

\item{modelParams}{A character vector of length <= 8 containing the names of the parameters for which traceplots will be made. If NULL, selects "plottable" attributes of object$bmList[[1]].}

\item{labels}{A character vector of the same length as modelParams, containing the y-axis labels that will be printed for each traceplot. The default is to simply use "modelParams".}

\item{title}{An optional title to be printed at the top of the traceplots.}

\item{file}{An optional location at which the traceplots will be saved. If NULL, no file is saved.}

\item{...}{additional plot arguments}
}
\value{
no return value
}
\description{
Given an object of class "mvBayes" from the mvBayes() function, traceplot() plots traceplots of user-specified variables from the Bayesian model fit, colored by basis.
}
\seealso{
See \link{mvBayes}
}
