% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{traceplot}
\alias{traceplot}
\title{Traceplots from a Bayesian Model Fit}
\usage{
traceplot(object, ...)
}
\arguments{
\item{object}{A fitted model object, e.g. of class "mvBayes".}

\item{...}{Additional arguments passed to methods.}
}
\value{
no return value
}
\description{
Generic function for plotting traceplots of the parameters of a
  fitted Bayesian model.
}
\seealso{
\link{traceplot.mvBayes}
}
