% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisSetup.R
\name{basisSetup}
\alias{basisSetup}
\title{Basis Expansion Calculation}
\usage{
basisSetup(
  Y,
  basisType = "pca",
  customBasis = NULL,
  nBasis = NULL,
  propVarExplained = 0.99,
  center = TRUE,
  scale = FALSE
)
}
\arguments{
\item{Y}{A matrix of dimension `c(n, nMV)`, where `n` is the number of observations of `nMV` response variables}

\item{basisType}{The type of basis functions to use. Options are `pca`, `pns`, `jfpca`, `jfpcah`, `splinet`, `bspline`, or `legendre`.}

\item{customBasis}{Optional user-provided basis of dimension `c(nBasis, nMV)`, only used if `basisType==custom`.}

\item{nBasis}{integer number of basis components to use (optional). By default, `propVarExplained` is used to choose nBasis.}

\item{propVarExplained}{Proportion (between 0 and 1) of variation to explain when choosing the number of basis components. Only used if `is.null(nBasis)`.}

\item{center}{logical: whether or to center `Y` before basis computations.}

\item{scale}{logical: whether to scale `Y` before basis computations.}
}
\value{
object of class `basisSetup` with plot method.
}
\description{
Calculates a basis expansion of a specified type.
}
\seealso{
Used by the main function \link{mvBayes}
}
\keyword{analysis,}
\keyword{basis}
\keyword{component}
\keyword{dimension}
\keyword{expansion,}
\keyword{principal}
\keyword{reduction}
