% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisSetup.R
\name{plot.basisSetup}
\alias{plot.basisSetup}
\title{Visualizing the Basis Decomposition}
\usage{
\method{plot}{basisSetup}(
  object,
  nBasis = NULL,
  propVarExplained = NULL,
  nPlot = NULL,
  idxMV = NULL,
  xscale = "linear",
  xlabel = "Multivariate Index",
  file = NULL,
  title = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class "basisSetup" containing information about a basis decomposition of a response matrix Y.}

\item{nBasis}{An integer specifying the number of basis functions to plot. If both nBasis and propVarExplained are NULL, object$nBasis is used, but anything <= object$nbasis is allowed.}

\item{propVarExplained}{Proportion (between 0 and 1) of variation to represent when choosing the number of bases to plot. It is only used if nBasis is NULL. If both are NULL, object$nBasis bases are plotted.}

\item{nPlot}{A positive integer specifying the number of samples to plot. Default is min(n, 1000), where n is nrow(Xtest) (or nrow(object$X) if Xtest is NULL).}

\item{file}{An optional location at which the plot will be saved. If NULL, no file is saved.}

\item{title}{An optional title to be printed at the top of the plot.}

\item{...}{additional plot arguments}
}
\value{
Top left: response Y. Top right: Y decomposed and colored by basis. Bottom left: Truncation error due to dimension reduction in the representation of Y. Bottom right: Percent variance explained by each basis.
}
\description{
Given an object of class "basisSetup" from the basisSetup() function, plot.basisSetup() visualizes the decomposition, colored by basis.
}
\seealso{
See \link{mvBayes}
}
\keyword{Bayesian}
\keyword{analysis}
\keyword{data}
\keyword{functional}
\keyword{modeling,}
\keyword{multivariate}
\keyword{regression}
